requirements.txt
python-telegram-bot
slackclient